Created a service for dealing with breakpoint sizes
This service has the ability to turn a size number or object into a set of bootstrap classes. What would be cool is if we could eventually use a service like this to make the css library pluggable. (Make it configurable what classes get set)

We'll probably need to create a separate module for just the components, in case consumers want to use them in a child module. (The providers only need to be included once in the application)