Implemented a card header
The header content is a derivitive of the data and the column set. If needed, we can provide the ability for the user to set a custom column template for the container header or the card headers. Clicking on the card header toggles the content to open or close. We'll have to implement some ability for only having one card open at a time.