Added a build step for building the library for distribution
For this, we'll build all files from the lib directory to a dist folder. Then we can set the dist index and index.d.ts files in the package.json, and consumers should be able to use this as a library. We'll probably want to upgrade this to enable AOT compilation in consumers. (I think you have to generate aot metadata files to support that)