Make the distribution compatible with systemjs
Currently we use systemjs-builder in our other projects. This doesn't allow for template urls but instead uses require to get the templates. We can do that here as webpack still seems to handle it okay, we just have to make sure the html files are available in the dist folder. To do this, I'll just add a simple node script that copies over the html files.