Added a pager component for showing buttons for changing the page
The card container will now input a 'current page' number and output paging events. It gives the pager a total count, page size (which is just the size of the data set), and page number, and the pager uses that to generate a list of paging buttons. These then just emit new page number events to the container, which then emits paging events with the number of items to get and the page being requested. We probably 'could' just emit the page number since the consumer controls the size anyway, but it seems better to emit them together for clarity, since we're essentially saying 'get the 3rd page of 10 items', or items 21-30.