Added a shrinkwrap file
This is supposed to make npm install the same package versions for other devs as it did for me. This should prevent them from having issues with broken package versions getting installed, while still allowing our package.json versions to be more flexible. For a library, this seems helpful since consumers may want to use this with slightly different versions of these packages, such as angular 2.5. As long as it still functions (which in most cases it should), we'd rather not prefent that or force them to get an update of this library first.