Created a service for handling default paging
This service will be used if paging is turned on in the simple card container. I wanted to make the service essentially just wrap a pure function, so we don't have to worry about keeping the paging state synced between the card container and the paging service. The only reason to make this a service at all is to allow consumers to override the provider with their own pager implementation. Otherwise these could just be exported functions.