Implemented default paging in the simple card container
To do this I'm implementing the concept of 'data filters' in the simple card container. This is essentially anything that we want to apply onto the data before it gets displayed on the page, and thus could be used for sorting as well (if we wanted to implement that). To enable paging, we can simply push the paging 'filter' into the filters list, and this will get applied before the data is displayed to the view. In the absence of any filters, the whole data set will be shown. (This is the normal behavior of lodash reduce)

Also, an advantage to having a single 'process' method that handles all processing is that we can reuse this anywhere we want to force the filtering to re-apply to update the view. At this point, we'll apply on init, anytime the data changes, and anytime the filter changes. The data will likely change once before init, but at that point there wont be any filters, so the 'process' operation should be cheap.