Publish the dist folder as the npm package
I'm still experimenting into the best ways to handle this. This is likely a temporary measure, since this package may eventually go to the ngx-template-seed format anyway. This is actually the way angular handles it for their internal packages. We have a parent package that installs our dependencies for development, and then a nested package for actually doing the publish.